refactor(api-file-manager): split monolithic abstractions.ts into per-file abstractions directory
Each asset delivery abstraction now lives in its own file under
abstractions/ (AssetRequestResolver, AssetResolver, AssetProcessor,
AssetOutputStrategy, AssetTransformationStrategy, AssetContentsReader,
AssetAuthorizer). The barrel abstractions.ts re-exports everything so
existing imports remain unchanged.
