+++ packages/api-file-manager/src/features/assetDelivery/abstractions/AssetResolver.ts
@@ -1,0 +1,15 @@
+import { createAbstraction } from "@webiny/feature/api";
+import type { AssetRequest as IAssetRequest } from "~/delivery/AssetDelivery/AssetRequest.js";
+import type { Asset as IAsset } from "~/delivery/AssetDelivery/Asset.js";
+
+export interface IAssetResolver {
+    resolve(request: IAssetRequest): Promise<IAsset | undefined>;
+}
+
+export const AssetResolver = createAbstraction<IAssetResolver>("AssetDelivery/AssetResolver");
+
+export namespace AssetResolver {
+    export type Interface = IAssetResolver;
+    export type Asset = IAsset;
+    export type Request = IAssetRequest;
+}
